+++ apps/api/src/surveys/mocks/user/users.ts
@@ -10,24 +10,7 @@
  * You should have received a copy of the GNU Affero General Public License along with this program. If not, see <https://www.gnu.org/licenses/>.
  */
 
-/* eslint-disable @typescript-eslint/no-unsafe-assignment */
 import { firstUsername, secondUsername } from './usernames';
-
-// export const firstMockUser = {
-//   email: 'first@example.com',
-//   username: firstUsername,
-//   roles: ['user'],
-//   mfaEnabled: false,
-//   isTotpSet: false,
-// };
-//
-// export const secondMockUser = {
-//   email: 'second@example.com',
-//   username: secondUsername,
-//   roles: ['user'],
-//   mfaEnabled: false,
-//   isTotpSet: false,
-// };
 
 export const firstMockUser = {
   firstName: 'first',
